Decouple Popover and NotificationPanel fragment
- Adds FloatingBox primitive
- Decouples styling completely
