+++ packages/css-framework/index.scss
@@ -89,6 +89,7 @@ $__states: null;
 
 // Objects
 @import "src/objects/divider";
+@import "src/objects/floating-box";
 @import "src/objects/hyperlink";
 @import "src/objects/media";
 @import "src/objects/wrapper";
