+++ packages/react-component-library/src/components/Popover/Popover.stories.tsx
@@ -1,43 +1,16 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 
-import {
-  POPOVER_ARROW_POSITION,
-  POPOVER_SCHEME,
-  POPOVER_PLACEMENT,
-} from './constants'
+import { FLOATING_BOX_PLACEMENT } from '../../primitives/FloatingBox'
+import { POPOVER_SCHEME } from './constants'
 
-import { Popover, PopoverOnHover } from '.'
+import { Popover } from '.'
 
 const stories = storiesOf('Popover', module)
 
-stories.add('Light', () => (
+stories.add('Default', () => (
   <Popover
-    top={25}
-    left={25}
-    height={200}
-    position={POPOVER_ARROW_POSITION.BOTTOM_LEFT}
-  >
-    <pre style={{ padding: '1rem' }}>This is some arbitrary JSX</pre>
-  </Popover>
-))
-
-stories.add('Dark', () => (
-  <Popover
-    top={25}
-    left={25}
-    height={200}
-    position={POPOVER_ARROW_POSITION.BOTTOM_LEFT}
-    scheme={POPOVER_SCHEME.DARK}
-  >
-    <pre style={{ padding: '1rem' }}>This is some arbitrary JSX</pre>
-  </Popover>
-))
-
-stories.add('On hover', () => (
-  <PopoverOnHover
-    placement={POPOVER_PLACEMENT.BELOW}
-    scheme={POPOVER_SCHEME.LIGHT}
+    placement={FLOATING_BOX_PLACEMENT.BELOW}
     popoverJSX={
       <pre style={{ padding: '1rem' }}>This is some arbitrary JSX</pre>
     }
@@ -51,5 +24,25 @@ stories.add('On hover', () => (
     >
       Hover on me!
     </div>
-  </PopoverOnHover>
+  </Popover>
+))
+
+stories.add('Dark', () => (
+  <Popover
+    placement={FLOATING_BOX_PLACEMENT.BELOW}
+    scheme={POPOVER_SCHEME.DARK}
+    popoverJSX={
+      <pre style={{ padding: '1rem' }}>This is some arbitrary JSX</pre>
+    }
+  >
+    <div
+      style={{
+        display: 'inline-block',
+        padding: '1rem',
+        backgroundColor: '#c9c9c9',
+      }}
+    >
+      Hover on me!
+    </div>
+  </Popover>
 ))
