+++ packages/react-component-library/src/components/Popover/index.ts
@@ -1,0 +1,1 @@
+export * from './Popover'
