--- packages/react-component-library/src/components/Popover/index.tsx
@@ -1,2 +1,0 @@
-export * from './Popover'
-export * from './PopoverOnHover'
