+++ packages/react-component-library/src/fragments/NotificationPanel/constants.ts
@@ -13,11 +13,11 @@ const NOTIFICATION_PLACEMENT_ARROW_POSITION_MAP = {
   [NOTIFICATION_PLACEMENT.RIGHT]: NOTIFICATION_ARROW_POSITION.LEFT_BOTTOM,
 }
 
-const POPOVER_WIDTH = 335
+const NOTIFICATION_CONTAINER_WIDTH = 335
 
 export {
   NOTIFICATION_ARROW_POSITION,
   NOTIFICATION_PLACEMENT,
   NOTIFICATION_PLACEMENT_ARROW_POSITION_MAP,
-  POPOVER_WIDTH,
+  NOTIFICATION_CONTAINER_WIDTH,
 }
