+++ packages/react-component-library/src/primitives/FloatingBox/constants.ts
@@ -1,0 +1,30 @@
+enum FLOATING_BOX_PLACEMENT {
+  ABOVE = 'above',
+  BELOW = 'below',
+  LEFT = 'left',
+  RIGHT = 'right'
+}
+
+enum FLOATING_BOX_ARROW_POSITION {
+  LEFT_BOTTOM = 'left_bottom',
+  LEFT_TOP = 'left_top',
+  RIGHT_BOTTOM = 'right_bottom',
+  RIGHT_TOP = 'right_top',
+  TOP_LEFT = 'top_left',
+  TOP_RIGHT = 'top_right',
+  BOTTOM_LEFT = 'bottom_left',
+  BOTTOM_RIGHT = 'bottom_right'
+}
+
+const FLOATING_BOX_PLACEMENT_ARROW_POSITION_MAP = {
+  [FLOATING_BOX_PLACEMENT.ABOVE]: FLOATING_BOX_ARROW_POSITION.BOTTOM_LEFT,
+  [FLOATING_BOX_PLACEMENT.BELOW]: FLOATING_BOX_ARROW_POSITION.TOP_LEFT,
+  [FLOATING_BOX_PLACEMENT.LEFT]: FLOATING_BOX_ARROW_POSITION.LEFT_BOTTOM,
+  [FLOATING_BOX_PLACEMENT.RIGHT]: FLOATING_BOX_ARROW_POSITION.RIGHT_BOTTOM,
+}
+
+export {
+  FLOATING_BOX_PLACEMENT,
+  FLOATING_BOX_ARROW_POSITION,
+  FLOATING_BOX_PLACEMENT_ARROW_POSITION_MAP,
+}
