+++ packages/react-component-library/src/primitives/FloatingBox/index.ts
@@ -1,0 +1,2 @@
+export * from './FloatingBox'
+export * from './constants'
