+++ packages/react-component-library/src/primitives/index.ts
@@ -1,0 +1,1 @@
+export { FloatingBox } from './FloatingBox'
